feat: removing PGPORT and PGHOST
Seven years ago, I added PGPORT and PGHOST to the exec-env script. I
have no idea why.  I didn't put why I did it into git because I didn't
understand the importance of git hygeine at the time.

[A ticket](https://github.com/smashedtoatoms/asdf-postgres/issues/68)
was opened to remove these variables, and though the idea of breaking
the contract bugs me, it doesn't bug me as much as the possibility that
I cargo-culted something in there that doesn't need to be in there that
is now causing problems.  I am going to remove it and see if anyone
notices.  If someone does though, I will revert this commit immediately.