fix: update docs and icu in actions
icu is an absolute tire fire on mac right now.  I disabled it in actions
because I don't have time to chase it down.  I updated the docs to
reflect the need for pkg-config because it mostly makes things easier.
Thank the postgres team and apple for making this a pain in the taint.